backend/requirements.txt
streamlit<=1.32.2
flask<=3.0.2
flask-cors<=5.0.0
langchain<=0.1.12
langchain-openai<=0.0.8
langchain_huggingface<=0.0.3
langchain_community<=0.2.7
openai<=1.14.0
SPARQLWrapper<=2.0.0
owlready2<=0.45
rdflib<=7.0.0
notebook<=7.1.2
faiss-cpu<=1.7.2
python-dotenv<=1.0.1
mistralai<=1.1.0
matplotlib<=3.9.2
seaborn<=0.12.2
weaviate-client<=4.10.4
chainlit<=2.1.0
